feat: Add default issue, PR, and discussion templates
We would like to standardize the structure of our issues and pull requests to ensure
that we are consistently capturing useful information and context. This aids in clear
communication by making sure that we are all structuring our work the same way. This
adds a set of templates that will automatically be included in all of the repositories
within the `mitodl` organization on GitHub.