config: Add org default Renovate configuration
In order to simplify the work of keeping our dependencies up to date we would like to
use the Renovate bot to manage automating the creation of PRs to update dependencies as
they are released. This is a base configuration that is in use for some of our
repositories and can serve as a good baseline for our other projects.